remove environment property from Store
The Store only uses the environment when it passes it to the reducer. But the Store
wraps the reducer in a closure that casts the environment to the proper type. Instead, that
closure can simply capture the environment as the proper type. Then the Store doesn't need
to, uh, store the environment in a property.